// ProjetWeb.Frontend/src/app/core/api/site/model/update-site-request.ts
/**
 * SiteManagement.API | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CreateCourtRequest } from './create-court-request';
import { CreatePlannedDayRequest } from './create-planned-day-request';


export interface UpdateSiteRequest { 
    name: string;
    closedDays: Array<string> | null;
    courts: Array<CreateCourtRequest> | null;
    schedule: Array<CreatePlannedDayRequest>;
}